debuginfo: Use `LocalRef` to simplify reference debuginfos
If the `LocalRef` is `LocalRef::Place`, we can refer to it directly,
because the local of place is an indirect pointer.
Such a statement is `_1 = &_2.1`.
If the `LocalRef` is `LocalRef::Operand`,
the `OperandRef` should provide the pointer of the reference.
Such a statement is `_1 = &(*_2).1`.

But there is a special case that hasn't been handled, scalar pairs like `(&[i32; 16], i32)`.